+++ modules/canvaskit/canvaskit_bindings.cpp
@@ -1160,9 +1160,12 @@ EMSCRIPTEN_BINDINGS(Skia) {
             // Otherwise, go with std::wstring and set UTF-32 encoding.
             return self.measureText(text.c_str(), text.length(), SkTextEncoding::kUTF8);
         }))
+        .function("setHinting", &SkFont::setHinting)
+        .function("setLinearMetrics", &SkFont::setLinearMetrics)
         .function("setScaleX", &SkFont::setScaleX)
         .function("setSize", &SkFont::setSize)
         .function("setSkewX", &SkFont::setSkewX)
+        .function("setSubpixel", &SkFont::setSubpixel)
         .function("setTypeface", &SkFont::setTypeface, allow_raw_pointers());
 
     class_<ShapedText>("ShapedText")
@@ -1701,11 +1704,19 @@ EMSCRIPTEN_BINDINGS(Skia) {
         .value("Round", SkPaint::Join::kRound_Join)
         .value("Bevel", SkPaint::Join::kBevel_Join);
 
+#ifndef SK_NO_FONTS
+    enum_<SkFontHinting>("FontHinting")
+        .value("None",   SkFontHinting::kNone)
+        .value("Slight", SkFontHinting::kSlight)
+        .value("Normal", SkFontHinting::kNormal)
+        .value("Full",   SkFontHinting::kFull);
+
     enum_<SkTextEncoding>("TextEncoding")
         .value("UTF8",    SkTextEncoding::kUTF8)
         .value("UTF16",   SkTextEncoding::kUTF16)
         .value("UTF32",   SkTextEncoding::kUTF32)
         .value("GlyphID", SkTextEncoding::kGlyphID);
+#endif
 
     enum_<SkTileMode>("TileMode")
         .value("Clamp",    SkTileMode::kClamp)